+++ offside/lib/MyPage/myTravelDetail.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:offside/TourSchedule/tourPlan.dart';
 import 'package:offside/data/api/tour_api.dart';
+import 'package:offside/data/model/tour_model.dart';
 import 'package:offside/data/view/team_info_view_model.dart';
 import 'package:offside/data/view/user_view_model.dart';
 
@@ -19,10 +20,12 @@ class _MyTravelDetail extends ConsumerState<MyTravelDetail> {
 
   @override
   Widget build(BuildContext context) {
+    print(widget.info);
     var size = MediaQuery.of(context).size;
     final teamInfoList = ref.watch(teamInfoViewModelProvider).teamInfoList;
     final user = ref.read(userViewModelProvider);
     var matchDate = widget.info.keys.first;
+
     var match = widget.info[matchDate]['match'];
     var tour = widget.info[matchDate]['tour'];
     return ListView(
@@ -34,22 +37,16 @@ class _MyTravelDetail extends ConsumerState<MyTravelDetail> {
               crossAxisAlignment: CrossAxisAlignment.center,
               children: [
                 const Padding(
-                  padding: const EdgeInsets.fromLTRB(0, 0, 0, 3.0),
-                  child: Icon(
-                    Icons.card_travel,
-                    size: 22,
-                    color: Color.fromRGBO(91, 143, 255, 1),
-                  ),
-                ),
-                SizedBox(
-                  width: size.width * 0.01,
-                ),
+                    padding: EdgeInsets.fromLTRB(0, 0, 0, 3.0),
+                    child: Icon(Icons.card_travel,
+                        size: 22, color: Color.fromRGBO(91, 143, 255, 1))),
+                SizedBox(width: size.width * 0.01),
                 Text('My Travel',
                     style: TextStyle(
                         fontWeight: FontWeight.w600,
                         fontSize: const AdaptiveTextSize()
                             .getadaptiveTextSize(context, 14),
-                        color: Color.fromRGBO(33, 58, 135, 1))),
+                        color: const Color.fromRGBO(33, 58, 135, 1)))
               ],
             )),
         Container(
@@ -67,16 +64,16 @@ class _MyTravelDetail extends ConsumerState<MyTravelDetail> {
                     children: [
                       Text('20${getDate(matchDate)}',
                           style: TextStyle(
-                              color: Color.fromARGB(255, 53, 53, 53),
+                              color: const Color.fromARGB(255, 53, 53, 53),
                               fontSize: const AdaptiveTextSize()
                                   .getadaptiveTextSize(context, 12))),
-                      SizedBox(height: 5),
+                      const SizedBox(height: 5),
                       Text(teamInfoList[match['home']].stadium,
                           style: TextStyle(
                               fontWeight: FontWeight.w500,
                               fontSize: const AdaptiveTextSize()
                                   .getadaptiveTextSize(context, 11),
-                              color: Color.fromARGB(255, 65, 65, 65)))
+                              color: const Color.fromARGB(255, 65, 65, 65)))
                     ],
                   ),
                   Row(
@@ -93,7 +90,7 @@ class _MyTravelDetail extends ConsumerState<MyTravelDetail> {
                             child: Image.network(
                                 teamInfoList[match['home']].logoImg),
                           ),
-                          SizedBox(height: 5),
+                          const SizedBox(height: 5),
                           Text(teamInfoList[match['home']].middleName,
                               style: TextStyle(
                                   fontWeight: FontWeight.w600,
@@ -118,7 +115,7 @@ class _MyTravelDetail extends ConsumerState<MyTravelDetail> {
                             child: Image.network(
                                 teamInfoList[match['away']].logoImg),
                           ),
-                          SizedBox(height: 5),
+                          const SizedBox(height: 5),
                           Text(teamInfoList[match['away']].middleName,
                               style: TextStyle(
                                   fontWeight: FontWeight.w600,
@@ -135,7 +132,7 @@ class _MyTravelDetail extends ConsumerState<MyTravelDetail> {
               crossAxisAlignment: CrossAxisAlignment.center,
               children: [
                 const Padding(
-                  padding: const EdgeInsets.fromLTRB(0, 0, 0, 3.0),
+                  padding: EdgeInsets.fromLTRB(0, 0, 0, 3.0),
                   child: Icon(
                     Icons.near_me,
                     size: 20,
@@ -150,11 +147,11 @@ class _MyTravelDetail extends ConsumerState<MyTravelDetail> {
                         fontWeight: FontWeight.w600,
                         fontSize: const AdaptiveTextSize()
                             .getadaptiveTextSize(context, 12),
-                        color: Color.fromRGBO(66, 66, 66, 1))),
+                        color: const Color.fromRGBO(66, 66, 66, 1))),
               ],
             )),
         Container(
-            height: tour.length * 80,
+            height: tour.length.toDouble() * 80,
             margin: const EdgeInsets.only(top: 5),
             child: ReorderableListView(
                 padding:
@@ -163,7 +160,7 @@ class _MyTravelDetail extends ConsumerState<MyTravelDetail> {
                   for (int index = 0; index < tour.length; index += 1)
                     InkWell(
                       key: Key('$index'),
-                      onLongPress: () {
+                      onDoubleTap: () {
                         showDialog(
                             context: context,
                             builder: (BuildContext context) {
@@ -226,7 +223,7 @@ class _MyTravelDetail extends ConsumerState<MyTravelDetail> {
                                       MainAxisAlignment.spaceBetween,
                                   crossAxisAlignment: CrossAxisAlignment.center,
                                   children: [
-                                    Text('${tour[index]['title']}',
+                                    Text(tour[index]['title'],
                                         style: TextStyle(
                                             fontSize: const AdaptiveTextSize()
                                                 .getadaptiveTextSize(
